python/requirements.txt
numpy
opencv-python
flask
cloudstorage
google-cloud-storage
